refactor: modularize BLoC structure for accounts and preferences
- Created separate files for AccountsEvent, AccountsState, PreferencesEvent, and PreferencesState to improve code organization and maintainability.
- Updated AccountsBloc and PreferencesBloc to utilize the new event and state classes.
- Enhanced SettingsScreen to include the new BLoC structure for better state management.
