+++ lib/features/settings/bloc/accounts_event.dart
@@ -1,0 +1,9 @@
+import 'package:equatable/equatable.dart';
+
+abstract class AccountsEvent extends Equatable {
+  @override
+  List<Object?> get props => [];
+}
+
+class SavingsTapped extends AccountsEvent {}
+class GiftsTapped extends AccountsEvent {}
