+++ lib/features/settings/bloc/accounts_state.dart
@@ -1,0 +1,8 @@
+import 'package:equatable/equatable.dart';
+
+class AccountsState extends Equatable {
+  final int savingsCount;
+  const AccountsState({required this.savingsCount});
+  @override
+  List<Object?> get props => [savingsCount];
+}
